Split watcher lifecycle activity helpers
