+++ docs/agent-maintenance/change-surfaces.md
@@ -131,7 +131,7 @@
   parsing), `src/services/discord/inflight.rs` (state file contract).
 - legacy_modules: none — relay routes are being consolidated, not replaced.
 - do_not_edit_without_migration_plan (giant-file):
-  - `src/services/discord/watchers/lifecycle.rs` (2318 lines — canonical
+  - `src/services/discord/watchers/lifecycle.rs` (2111 lines — canonical
     lifecycle extraction surface from #1435; split further before adding new
     lifecycle behavior; #3016 phase-5b2 dropped the `mailbox_finalize_owed`
     construction from the watcher-spawn handle; #3718 moved runtime mtime
@@ -139,7 +139,9 @@
     keeps lifecycle below its frozen ratchet; -6 from #3736 removing legacy
     remote-profile restore plumbing while remote SSH is disabled; ±0 from
     #3815 moving direct Codex TUI resume restore helpers into
-    `watchers/codex_tui_restore.rs` while adding the restore branch).
+    `watchers/codex_tui_restore.rs` while adding the restore branch; -207 from
+    #3840 moving heartbeat/activity helpers into
+    `watchers/lifecycle/activity.rs`).
   - `src/services/discord/tmux.rs` (2039 lines after #2558 dead-code sweep;
     +6 from #3818 sanitizing restored/orphan subagent-notification placeholders;
     +1 from #3384 restored-seed undelivered-body discard guard;
